tambahan dan perbaikan yang eror

+++ pengelola/delete_semprop_diadmin.php
@@ -35,12 +35,19 @@ if($_SESSION['status'] == "login"){
     <style type="text/css" href="../css/tombol_penjadwalan.css"></style>
 </head>
 <body>
-
-
-    <br>
+<br>
 
     <h2 align="center">DATA SEMINAR PROPOSAL</h2>
+    <table width='90%'>
+        <th>
+          <td align='right' > <a target="_blank" href="export_excel_semprop.php">EXPORT KE EXCEL</a><br></td>
+          
+          </th>
+      </table>
 <br>
+
+   
+
 <table align="center">
 <form name="pencarian" method="POST" action = "hasil_cari_pengunguman_diadmin.php" ">            
       
@@ -90,16 +97,20 @@ if($_SESSION['status'] == "login"){
       <br>
 
 
-  <table border='1' align='center' width='80%'' height='30%'>
+  <table border='1' align='center' width='95%'' height='30%'>
     <tr align='center' bgcolor='#D3D3D3'>
       <th height='50'>Nim</th>
       <th height='50' >Nama</th>
       <th height='50'>Nilai Proses pembimbing</th>
        <th height='50'>Nilai ujian pembimbing</th>
         <th height='50'>Nilai ujian penguji</th>
+         <th height='50'>Rata-rata</th>
+          <th height='50'>Grade</th>
+
       <th height='50'>Status</th>
       <th height='50'>Action</th>
     </tr>
+
 
 
 
@@ -111,22 +122,33 @@ $akses->DeleteDataSemprop($nim);
 
 foreach ($akses->LihatPengumumanNilaiDanStatusSemuaMahasiswa() as $key) {
         
-        
-           
-        
+      #DIBUAT OLHE IBRAHIM
+ $rata_rata=round(($key['nilai_proses_pembimbing']+$key['nilai_ujian_pembimbing']+$key['nilai_ujian_penguji'])/3,2);
+if($rata_rata>-1 && $rata_rata<=1) $grade='E';
+  else if($rata_rata>0 && $rata_rata<=40) $grade='D';
+  else if($rata_rata>40 && $rata_rata<=43.75) $grade='D+';
+  else if($rata_rata>43.75 && $rata_rata<=51.25) $grade='C-';
+  else if($rata_rata>51.25 && $rata_rata<=55) $grade='C';
+  else if($rata_rata>55 && $rata_rata<=57.5) $grade='C+';
+  else if($rata_rata>57.5 && $rata_rata<=62.5) $grade='B-';
+  else if($rata_rata>62.5 && $rata_rata<=65) $grade='B';
+  else if($rata_rata>65 && $rata_rata<=68.75) $grade='B+';
+  else if($rata_rata>68.75 && $rata_rata<=76.25) $grade='A-';
+  else if($rata_rata>76.25 && $rata_rata<=100) $grade='A';
+  else $grade='nilai tidak tersedia'; 
 
-
-        echo"
-       
-        
-
+  if($rata_rata>51.25) $status='lulus';
+  else($status='tidak_lulus')  ;  
+        echo "
         <tr>
           <td align='center'>$key[nim]</td>
           <td align='center'>$key[nama_mhs]</td>
           <td align='center'>$key[nilai_proses_pembimbing]</td>
            <td align='center'>$key[nilai_ujian_pembimbing]</td>
             <td align='center'>$key[nilai_ujian_penguji]</td>
-          <td align='center'>$key[status]</td>
+          <td align='center'>$rata_rata</td>
+          <td align='center'>$grade</td>
+          <td align='center'>$status</td>
           <td align='center'><a href='update_semrop_diadmin.php?nim=$key[nim]' role='button' class='btn btn-outline-primary'>UPDATE</a>
           <a href='delete_semprop_diadmin.php?nim=$key[nim]' role='button' class='btn btn-outline-primary'>DELETE</a></td>
           </tr>
